requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.3
Django==5.1
django-adminlte3==0.1.6
django-bootstrap4==24.3
django-crispy-forms==2.3
django-pictures==1.3.3
django-stdimage==6.0.2
googletrans==3.0.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.23.0
hyperframe==5.2.0
idna==3.7
joblib==1.4.2
model-bakery==1.20.0
model-mommy==2.0.0
nltk==3.9.1
packaging==24.1
pillow==10.4.0
python-decouple==3.8
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
style==1.1.0
textblob==0.18.0.post0
tqdm==4.66.5
tzdata==2024.1
update==0.0.1
urllib3==2.2.3
whitenoise==6.7.0
